Added a usable showLong to the repl
- There are quite a few hacky bits added to the type checker now, which were
  necessary to resolve boxed-vs-unboxed arguments to functions. This should be
  temporary, though, as even the hacks don't work particularly well, and we need
  to revisit how we do some things to make them unnecessary.

- Also fixed a tiny test suite bug.

+++ src/Ermine/Parser/Literal.hs
@@ -18,6 +18,6 @@ import Text.Parser.Token
 
 -- | Parse a 'literal' number, string or character.
 literal :: (Monad m, TokenParsing m) => m Literal
-literal = either (Int . fromIntegral) Double <$> naturalOrDouble
+literal = either (Long . fromIntegral) Double <$> naturalOrDouble
       <|> String <$> stringLiteral
       <|> Char <$> charLiteral
